update runmode code
